Allow Supermarket to use S3
The current community cookbooks are hosted on S3 so Supermarket
must be able to support S3 for cookbook storage. Now if both a S3 bucket
is set paperclip will be configured to use S3 otherwise it will use
local storage. This is done so the application is not dependent upon S3.